Remove encryption from stored policies JSON since there's no secrets in there anymore
